refactor: replace Button with ButtonCVA
