+++ src/app/(auth)/sign-up/page.tsx
@@ -26,8 +26,8 @@ export default function Component() {
   });
 
   return (
-    <div className="w-full max-w-md px-8 py-6 bg-white rounded-xl shadow-md space-y-6">
-      <h2 className="text-center text-3xl font-extrabold text-gray-900">
+    <div className="w-full max-w-md px-8 py-6 space-y-6 bg-white shadow-md rounded-xl">
+      <h2 className="text-3xl font-extrabold text-center text-gray-900">
         Register
       </h2>
       <form className="space-y-6">
@@ -63,12 +63,8 @@ export default function Component() {
             maxLength={30}
           />
         </div>
-        <Button
-          title="title"
-          className="w-full bg-blue-500 text-white rounded-md py-2"
-          type="submit"
-        >
-          Login
+        <Button size="lg" className="w-full" type="submit">
+          Join now
         </Button>
       </form>
     </div>
