+++ src/app/(main)/directory/components/EditCell.tsx
@@ -1,6 +1,5 @@
 import { PencilSquareIcon } from "@heroicons/react/24/solid";
 import type { TeamMember } from "./fixtures/MyTeam";
-import Button from "@/components/Button";
 
 interface EditCellProps {
   teamMember: TeamMember;
@@ -17,12 +16,12 @@ export default function EditCell({ teamMember, currentUserId }: EditCellProps) {
     >
       {teamMember.averageHour === 0 ? "Add hours" : teamMember.averageHour}
       {teamMember.id === currentUserId && (
-        <Button
+        <button
           title="edit"
-          customClassName="pl-2 pr-1 h-full rounded-l-none rounded-r-md p-0 min-h-0 text-sm font-medium text-base-300 bg-transparent border-transparent hover:bg-transparent hover:border-transparent"
+          className="h-full min-h-0 p-0 pl-2 pr-[10px] text-sm font-medium bg-transparent border-transparent rounded-l-none rounded-r-md text-base-300 hover:bg-transparent hover:border-transparent"
         >
           <PencilSquareIcon className="w-4 h-4 text-base-300" />
-        </Button>
+        </button>
       )}
     </div>
   );
