+++ src/app/(main)/ideation/components/CreateIdeationContainer.tsx
@@ -6,15 +6,12 @@ function CreateIdeationContainer() {
     <div className="card w-full min-h-[190px] max-[1919px]:min-h-[280px] bg-base-100 flex flex-row items-center px-10">
       <div className="flex flex-col justify-between h-full py-10 min-[1920px]:hidden">
         <VoteDescriptionCard />
-        <Button
-          title="Create Project"
-          customClassName="max-[1920px]:w-[160px] btn-primary text-base-300 capitalize"
-        >
+        <Button size="lg" className="max-[1920px]:w-[160px]">
           Create Project
         </Button>
       </div>
       <VoteDescriptionCard className="hidden min-[1920px]:block" />
-      <section className="card-body gap-y-7 px-20">
+      <section className="px-20 card-body gap-y-7">
         <h2 className="text-xl font-semibold text-base-300">
           What is your Voyage project idea & vision?
         </h2>
@@ -25,8 +22,8 @@ function CreateIdeationContainer() {
         </p>
       </section>
       <Button
-        title="Create Project"
-        customClassName="max-w-[200px] w-full btn-primary text-base-300 capitalize hidden min-[1920px]:block"
+        size="lg"
+        className="max-w-[200px] w-full hidden min-[1920px]:block"
       >
         Create Project
       </Button>
