+++ src/app/(main)/ideation/components/VoteCard.tsx
@@ -17,16 +17,12 @@ function VoteCard({ users, voted, className }: VoteCardProps) {
         <h2 className="text-xl font-semibold text-base-300">{`Vote${
           users.length > 1 ? "s" : ""
         }`}</h2>
-        <div className="avatar-group -space-x-2 w-full">
+        <div className="w-full -space-x-2 avatar-group">
           {users.map((user) => (
             <Avatar width={24} height={24} key={user} />
           ))}
         </div>
-        <Button
-          title="Vote"
-          customClassName="w-full btn-primary text-base-300 disabled:bg-primary-focus disabled:text-base-200 capitalize"
-          disabled={voted}
-        >
+        <Button className="w-full" disabled={voted}>
           {voted ? "Voted" : "Vote"}
         </Button>
       </section>
