+++ src/app/(main)/tech-stack/components/TechStackCard.tsx
@@ -12,11 +12,11 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
   return (
     <div className="card min-w-[400px] sm:w-96 text-base-300 bg-base-200 rounded-lg px-6 py-5">
       <div className="flex flex-row justify-between">
-        <h3 className="text-xl font-semibold text-base-300 self-center">
+        <h3 className="self-center text-xl font-semibold text-base-300">
           {title}
         </h3>
       </div>
-      <div className="h-40 overflow-y-auto mt-6 pt-1">
+      <div className="h-40 pt-1 mt-6 overflow-y-auto">
         <ul className="text-base-300">
           {data.map((element) => (
             <li
@@ -24,7 +24,7 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
               key={element.id}
             >
               {element.value}
-              <div className="avatar-group -space-x-2 absolute left-28">
+              <div className="absolute -space-x-2 avatar-group left-28">
                 {element.users.map((user) => (
                   <Avatar
                     key={`${element.id}-${user}`}
@@ -33,20 +33,14 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
                   />
                 ))}
               </div>
-              <Button
-                title={title}
-                customClassName="capitalize w-[62px] h-[32px] p-0 min-h-full text-xs font-medium text-base-300 bg-primary-content border-transparent mr-4 rounded-[32px] hover:bg-primary hover:border-transparent gap-x-0"
-              >
+              <button className="capitalize w-[62px] h-[32px] p-0 min-h-full text-xs font-medium text-base-300 bg-primary-content border-transparent mr-4 rounded-[32px] hover:bg-primary hover:border-transparent gap-x-0">
                 Vote
-              </Button>
+              </button>
             </li>
           ))}
         </ul>
       </div>
-      <Button
-        title={`add ${title}`}
-        customClassName="mt-6 capitalize w-full h-[42px] p-0 min-h-full text-xs font-medium text-base-300 bg-secondary border-transparent flex justify-start pl-5 items-center hover:bg-secondary hover:border-transparent"
-      >
+      <Button variant="secondary" className="justify-start w-full">
         <PlusCircleIcon className="h-[18px] w-[18px] text-base-300" />
         Add Tech Stack
       </Button>
